requirements.txt
# PrajnaAI — Complete Requirements
# All packages work offline once downloaded
# Install: pip install -r requirements.txt

# ── Core Scientific Stack ──────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ── Machine Learning ──────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
shap>=0.42.0
imbalanced-learn>=0.11.0
joblib>=1.3.0

# ── Deep Learning (CPU PyTorch) ───────────────────────────────────────────
# Install CPU-only torch first: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── NLP ───────────────────────────────────────────────────────────────────
nltk>=3.8.1
spacy>=3.6.0
transformers>=4.35.0       # for tokenizers and small models
sentence-transformers>=2.2.2
tokenizers>=0.14.0

# ── LLM / RAG / Agents ───────────────────────────────────────────────────
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
langgraph>=0.1.0
chromadb>=0.4.22
ollama>=0.1.7

# ── Computer Vision ───────────────────────────────────────────────────────
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── Data I/O ──────────────────────────────────────────────────────────────
openpyxl>=3.1.0
pyarrow>=13.0.0
pdfplumber>=0.10.0
pypdf>=3.17.0

# ── UI / Visualization ────────────────────────────────────────────────────
streamlit>=1.28.0
streamlit-chat>=0.1.1
altair>=5.1.0
rich>=13.6.0
tqdm>=4.66.0

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.4.0
httpx>=0.25.0
tenacity>=8.2.0
loguru>=0.7.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
